Rip out old plumbing to rely solely on click for verdi
Now that all `verdi` commands have been migrated to use `click`,
the homebrew cli infrastructure can be ripped out and the entire
stack is now powered by `click`. There were three commands, `setup`
`quicksetup` and `run` that did not yet have their own command.
They have been fully clickerized and hooked up now.
