+++ aiida/cmdline/commands/cmd_computer.py
@@ -10,23 +10,25 @@
 """`verdi computer` commands"""
 import sys
 import click
-from click_plugins import with_plugins
 
-from aiida.cmdline.baseclass import VerdiCommandWithSubcommands
-from aiida.backends.utils import load_dbenv, is_dbenv_loaded
 from aiida.common.exceptions import ValidationError
 from aiida.common.utils import escape_for_bash
-from aiida.cmdline.commands import verdi, verdi_computer, ensure_scripts
+from aiida.cmdline.commands import verdi
 from aiida.cmdline.params import options, arguments
-from aiida.cmdline.utils import echo
-from aiida.cmdline.utils.decorators import with_dbenv
 from aiida.cmdline.params import types
 from aiida.cmdline.params.options.interactive import InteractiveOption
-from aiida.cmdline.params.types import (
-    ShebangParamType, MpirunCommandParamType, NonemptyStringParamType)
+from aiida.cmdline.params.types import ShebangParamType, MpirunCommandParamType, NonemptyStringParamType
+from aiida.cmdline.utils import echo
+from aiida.cmdline.utils.decorators import with_dbenv
+from aiida.cmdline.utils.multi_line_input import ensure_scripts
 from aiida.control.computer import ComputerBuilder, get_computer_configuration
 from aiida.plugins.entry_point import get_entry_points
-from aiida import transport
+
+
+@verdi.group('computer')
+def verdi_computer():
+    """Setup and manage computers."""
+    pass
 
 
 def get_computer_names():
@@ -40,6 +42,7 @@ def get_computer_names():
         return zip(*qb.all())[0]
     else:
         return []
+
 
 @with_dbenv()
 def get_computer(name):
@@ -55,6 +58,7 @@ def prompt_for_computer_configuration(computer):
     import inspect, readline
     from aiida.orm.computer import Computer
     from aiida.common.exceptions import ValidationError
+
 
 def shouldcall_default_mpiprocs_per_machine(ctx):
     """
@@ -81,7 +85,8 @@ def shouldcall_default_mpiprocs_per_machine(ctx):
 
     return JobResourceClass.accepts_default_mpiprocs_per_machine()
 
-def _computer_test_get_jobs(transport,scheduler,authinfo):
+
+def _computer_test_get_jobs(transport, scheduler, authinfo):
     """
     Internal test to check if it is possible to check the queue state.
 
@@ -98,6 +103,7 @@ def _computer_test_get_jobs(transport,scheduler,authinfo):
     #    print jid, data['submission_time'], data['dispatch_time'], data['job_state']
     echo.echo("  `-> OK, {} jobs found in the queue.".format(len(found_jobs)))
     return True
+
 
 def _computer_create_temp_file(transport, scheduler, authinfo):
     """
@@ -169,61 +175,70 @@ def _computer_create_temp_file(transport, scheduler, authinfo):
     echo.echo("  [Deleted successfully]")
     return True
 
+
 @verdi_computer.command('setup')
 @click.pass_context
 @options.LABEL(prompt='Computer label', cls=InteractiveOption, required=True, type=NonemptyStringParamType())
-@options.HOSTNAME(prompt='Hostname', cls=InteractiveOption, required=True,
+@options.HOSTNAME(
+    prompt='Hostname',
+    cls=InteractiveOption,
+    required=True,
     help="The fully qualified host-name of this computer; for local transports, use 'localhost'")
-@options.DESCRIPTION(prompt='Description', cls=InteractiveOption,
-                     help="A human-readable description of this computer")
-@click.option('-e/-d', '--enabled/--disabled', is_flag=True, default=True,
+@options.DESCRIPTION(prompt='Description', cls=InteractiveOption, help="A human-readable description of this computer")
+@click.option(
+    '-e/-d',
+    '--enabled/--disabled',
+    is_flag=True,
+    default=True,
     help='if created with the disabled flag, calculations '
-         'associated with it will not be submitted until when it is '
-         're-enabled',
+    'associated with it will not be submitted until when it is '
+    're-enabled',
     prompt="Enable the computer?",
     cls=InteractiveOption,
     # IMPORTANT! Do not specify explicitly type=click.BOOL,
     # Otherwise you would not get a default value when prompting
- )
+)
 @options.TRANSPORT(prompt="Transport plugin", cls=InteractiveOption)
 @options.SCHEDULER(prompt="Scheduler plugin", cls=InteractiveOption)
 @click.option(
-   '--shebang',
+    '--shebang',
     prompt='Shebang line (first line of each script, starting with #!)',
     default="#!/bin/bash",
     cls=InteractiveOption,
     help='this line specifies the first line of the submission script for this computer',
     type=ShebangParamType())
 @click.option(
-    '-w', '--work-dir',
+    '-w',
+    '--work-dir',
     prompt='work directory on the computer',
     default="/scratch/{username}/aiida/",
     cls=InteractiveOption,
     help="The absolute path of the directory on the computer where AiiDA will "
-         "run the calculations (typically, the scratch of the computer). You "
-         "can use the {username} replacement, that will be replaced by your "
-         "username on the remote computer")
+    "run the calculations (typically, the scratch of the computer). You "
+    "can use the {username} replacement, that will be replaced by your "
+    "username on the remote computer")
 @click.option(
-    '-m', '--mpirun-command',
+    '-m',
+    '--mpirun-command',
     prompt="mpirun command",
     default="mpirun -np {tot_num_mpiprocs}",
     cls=InteractiveOption,
     help="The mpirun command needed on the cluster to run parallel MPI "
-         "programs. You can use the {tot_num_mpiprocs} replacement, that will be "
-         "replaced by the total number of cpus, or the other scheduler-dependent "
-         "replacement fields (see the scheduler docs for more information)",
+    "programs. You can use the {tot_num_mpiprocs} replacement, that will be "
+    "replaced by the total number of cpus, or the other scheduler-dependent "
+    "replacement fields (see the scheduler docs for more information)",
     type=MpirunCommandParamType())
 @click.option(
     '--mpiprocs-per-machine',
     prompt="default number of CPUs per machine",
     cls=InteractiveOption,
     help="Enter here the default number of MPI processes per machine (node) that "
-         "should be used if nothing is otherwise specified. Pass the digit 0 "
-         "if you do not want to provide a default value.",
+    "should be used if nothing is otherwise specified. Pass the digit 0 "
+    "if you do not want to provide a default value.",
     prompt_fn=shouldcall_default_mpiprocs_per_machine,
     required_fn=False,
     type=click.INT,
-) # Note: this can still be passed from the command line in non-interactive mode
+)  # Note: this can still be passed from the command line in non-interactive mode
 @options.PREPEND_TEXT()
 @options.APPEND_TEXT()
 @options.NON_INTERACTIVE()
@@ -235,9 +250,9 @@ def setup_computer(ctx, non_interactive, **kwargs):
 
     if kwargs['label'] in get_computer_names():
         echo.echo_critical('A computer called {} already exists.\n'
-            'Use "verdi computer update" to update it, and be '
-            'careful if you really want to modify a database '
-            'entry!'.format(kwargs['label']))
+                           'Use "verdi computer update" to update it, and be '
+                           'careful if you really want to modify a database '
+                           'entry!'.format(kwargs['label']))
 
     if not non_interactive:
         pre, post = ensure_scripts(kwargs.pop('prepend_text', ''), kwargs.pop('append_text', ''), kwargs)
@@ -252,7 +267,6 @@ def setup_computer(ctx, non_interactive, **kwargs):
         computer = computer_builder.new()
     except (ComputerBuilder.ComputerValidationError, ValidationError) as e:
         echo.echo_critical('{}: {}'.format(type(e).__name__, e))
-
 
     try:
         computer.store()
@@ -559,7 +573,6 @@ def computer_test(user, print_traceback, computer):
         succeeded = False
 
 
-
 @verdi_computer.command('delete')
 @arguments.COMPUTER()
 @with_dbenv()
@@ -583,12 +596,11 @@ def computer_delete(computer):
     echo.echo_success("Computer '{}' deleted.".format(compname))
 
 
-
-# ~ @with_plugins(get_entry_points('aiida.cmdline.computer.configure'))
 @verdi_computer.group('configure')
 def computer_configure():
     """Configure a computer with one of the available transport types."""
     pass
+
 
 @computer_configure.command('show')
 @click.option('--current/--defaults')
@@ -599,16 +611,20 @@ def computer_config_show(computer, user, current, as_option_string):
     """Show the current or default configuration for COMPUTER."""
     import tabulate
     from pprint import pformat
+    from aiida.transport import cli as transport_cli
     config = {}
     table = []
 
     transport_cls = computer.get_transport_class()
-    option_list = [param for param in transport.cli.create_configure_cmd(computer.get_transport_type()).params if isinstance(param, click.core.Option)]
+    option_list = [
+        param for param in transport_cli.create_configure_cmd(computer.get_transport_type()).params
+        if isinstance(param, click.core.Option)
+    ]
     option_list = [option for option in option_list if option.name in transport_cls.get_valid_auth_params()]
     if current:
         config = get_computer_configuration(computer, user)
     else:
-        config = {option.name: transport.cli.transport_option_default(option.name, computer) for option in option_list}
+        config = {option.name: transport_cli.transport_option_default(option.name, computer) for option in option_list}
 
     option_items = []
     if as_option_string:
@@ -616,7 +632,8 @@ def computer_config_show(computer, user, current, as_option_string):
             t_opt = transport_cls.auth_options[option.name]
             if config.get(option.name) or config.get(option.name) is False:
                 if t_opt.get('switch'):
-                    option_value = option.opts[-1] if config.get(option.name) else '--no-{}'.format(option.name.replace('_', '-'))
+                    option_value = option.opts[-1] if config.get(option.name) else '--no-{}'.format(
+                        option.name.replace('_', '-'))
                 elif t_opt.get('is_flag'):
                     is_default = config.get(option.name) == transport_option_default(option.name, computer)
                     option_value = option.opts[-1] if is_default else ''
@@ -629,102 +646,7 @@ def computer_config_show(computer, user, current, as_option_string):
         table = [('* ' + name, config[name]) for name in transport_cls.get_valid_auth_params()]
         echo.echo(tabulate.tabulate(table, tablefmt='plain'))
 
+
 for ep in get_entry_points('aiida.transports'):
-    computer_configure.add_command(transport.cli.create_configure_cmd(ep.name))
-
-class Computer(VerdiCommandWithSubcommands):
-    """
-    Setup and manage computers to be used
-
-    This command allows to list, add, modify and configure computers.
-    """
-
-    def __init__(self):
-        """
-        A dictionary with valid commands and functions to be called.
-        """
-        super(Computer, self).__init__()
-
-        self.valid_subcommands = {
-            'list': (verdi, self.complete_none),
-            'show': (verdi, self.complete_computers),
-            'setup': (verdi, self.complete_none),
-            'update': (self.computer_update, self.complete_computers),
-            'enable': (verdi, self.complete_computers),
-            'disable': (verdi, self.complete_computers),
-            'rename': (verdi, self.complete_computers),
-            'test': (verdi, self.complete_computers),
-            'delete': (verdi, self.complete_computers),
-            'configure': (verdi, self.complete_none),
-        }
-
-    def complete_computers(self, subargs_idx, subargs):
-        if not is_dbenv_loaded():
-            load_dbenv()
-        computer_names = get_computer_names()
-        print computer_names
-        return "\n".join(computer_names)
-
-    def computer_update(self, *args):
-        """
-        Update an existing computer
-        """
-        import argparse
-        from aiida.common.exceptions import NotExistent
-
-        if not is_dbenv_loaded():
-            load_dbenv()
-
-        from aiida.orm.computer import Computer
-
-        parser = argparse.ArgumentParser(prog=self.get_full_command_name(), description='Update a computer')
-        # The default states are those that are shown if no option is given
-        parser.add_argument('computer_name', help="The name of the computer")
-        parsed_args = parser.parse_args(args)
-        computer_name = parsed_args.computer_name
-
-        try:
-            computer = Computer.get(computer_name)
-        except NotExistent:
-            print "No computer {} was found".format(computer_name)
-            sys.exit(1)
-
-        calculation_on_computer = computer.get_calculations_on_computer()
-
-        if calculation_on_computer:
-            # Note: this is an artificial comment.
-            # If you comment the following lines, you will be able to overwrite
-            # the old computer anyway, at your own risk.
-            print "You cannot modify a computer, after you run some calculations on it."
-            print "Disable this computer and set up a new one."
-            sys.exit(1)
-
-        print "*" * 75
-        print "WARNING! Modifying existing computer with name '{}'".format(computer_name)
-        print "Are you sure you want to continue? The UUID will remain the same!"
-        print "Continue only if you know what you are doing."
-        print "If you just want to rename a computer, use the 'verdi computer rename'"
-        print "command. In most cases, it is better to create a new computer."
-        print "Moreover, if you change the transport, you must also reconfigure"
-        print "each computer for each user!"
-        print "*" * 75
-        print "Press [Enter] to continue, or [Ctrl]+C to exit."
-        raw_input()
-
-        prompt_for_computer_configuration(computer)
-
-        try:
-            computer.store()
-        except ValidationError as e:
-            print "Unable to store the computer: {}. Exiting...".format(e.message)
-            sys.exit(1)
-
-        print "Computer '{}' successfully updated.".format(computer_name)
-        print "pk: {}, uuid: {}".format(computer.pk, computer.uuid)
-        print "(Note: machine_dependent transport parameters cannot be set via "
-        print "the command-line interface at the moment)"
-
-        print "OK"
-        pass
-
-
+    from aiida.transport import cli as transport_cli
+    computer_configure.add_command(transport_cli.create_configure_cmd(ep.name))
