+++ aiida/cmdline/commands/cmd_data/__init__.py
@@ -7,63 +7,25 @@
 # For further information on the license, see the LICENSE.txt file        #
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
-
+"""The `verdi data` command line interface."""
 import click
 
-from aiida.cmdline.baseclass import VerdiCommandWithSubcommands
-from aiida.cmdline.commands import verdi, verdi_data
+from aiida.cmdline.commands import verdi
 from aiida.cmdline.utils import decorators, echo
+from aiida.cmdline.utils.pluginable import Pluginable
 
 
-class Data(VerdiCommandWithSubcommands):
-    """
-    Setup and manage data specific types
-
-    There is a list of subcommands for managing specific types of data.
-    For instance, 'data upf' manages pseudopotentials in the UPF format.
-    """
-
-    def __init__(self):
-        super(Data, self).__init__()
-
-        self.valid_subcommands = {
-            'upf': (self.cli, self.complete_none),
-            'structure': (self.cli, self.complete_none),
-            'bands': (self.cli, self.complete_none),
-            'cif': (self.cli, self.complete_none),
-            'trajectory': (self.cli, self.complete_none),
-            'parameter': (self.cli, self.complete_none),
-            'array': (self.cli, self.complete_none),
-            'remote': (self.cli, self.complete_none),
-            'plugins': (self.cli, self.complete_plugins),
-        }
-
-    @staticmethod
-    def cli(*args):  # pylint: disable=unused-argument
-        verdi()  # pylint: disable=no-value-for-parameter
-
-    @staticmethod
-    def _ctx(args, info_name='verdi data', **kwargs):
-        return verdi_data.make_context(info_name, list(args), **kwargs)
-
-    @staticmethod
-    @decorators.with_dbenv()
-    def complete_plugins(subargs_idx, subargs):
-        """Return the list of plugins registered under the 'data' category."""
-        from aiida.plugins.entry_point import get_entry_point_names
-
-        other_subargs = subargs[:subargs_idx] + subargs[subargs_idx + 1:]
-        return_plugins = [e for e in get_entry_point_names('aiida.data') if e not in other_subargs]
-        return '\n'.join(return_plugins)
+@verdi.group('data', entry_point_group='aiida.cmdline.data', cls=Pluginable)
+def verdi_data():
+    """Inspect, create and manage data nodes."""
+    pass
 
 
 @verdi_data.command('plugins')
 @click.argument('entry_point', type=click.STRING, required=False)
 @decorators.with_dbenv()
 def data_plugins(entry_point):
-    """
-    Print a list of registered data plugins or details of a specific data plugin
-    """
+    """Print a list of registered data plugins or details of a specific data plugin."""
     from aiida.common.exceptions import LoadingPluginFailed, MissingPluginError
     from aiida.plugins.entry_point import get_entry_point_names, load_entry_point
 
@@ -79,8 +41,8 @@ def data_plugins(entry_point):
         entry_points = get_entry_point_names('aiida.data')
         if entry_points:
             echo.echo('Registered data entry points:')
-            for entry_point in entry_points:
-                echo.echo("* {}".format(entry_point))
+            for registered_entry_point in entry_points:
+                echo.echo("* {}".format(registered_entry_point))
 
             echo.echo('')
             echo.echo_info('Pass the entry point as an argument to display detailed information')
@@ -88,7 +50,8 @@ def data_plugins(entry_point):
             echo.echo_error('No data plugins found')
 
 
-# import to populate the click subcommands
+# Import to populate the `verdi data` sub commands
+# pylint: disable=wrong-import-position
 from aiida.cmdline.commands.cmd_data import (
     cmd_array,
     cmd_bands,
@@ -98,4 +61,4 @@ from aiida.cmdline.commands.cmd_data import (
     cmd_structure,
     cmd_trajectory,
     cmd_upf,
-    )
+)
