+++ aiida/cmdline/commands/cmd_data/cmd_trajectory.py
@@ -11,7 +11,7 @@
 This allows to manage TrajectoryData objects from command line.
 """
 import click
-from aiida.cmdline.commands import verdi_data
+from aiida.cmdline.commands.cmd_data import verdi_data
 from aiida.cmdline.commands.cmd_data.cmd_export import _export, export_options
 from aiida.cmdline.commands.cmd_data.cmd_list import _list, list_options
 from aiida.cmdline.commands.cmd_data.cmd_show import show_options
@@ -20,10 +20,8 @@ from aiida.cmdline.utils import echo
 from aiida.backends.utils import load_dbenv, is_dbenv_loaded
 
 
-# pylint: disable=unused-argument
 @verdi_data.group('trajectory')
-@click.pass_context
-def trajectory(ctx):
+def trajectory():
     """
     View and manipulate TrajectoryData instances.
     """
