+++ aiida/cmdline/commands/cmd_node.py
@@ -14,56 +14,21 @@ import sys
 import click
 import tabulate
 
-from aiida.backends.utils import load_dbenv, is_dbenv_loaded
-from aiida.cmdline import delayed_load_node as load_node
-from aiida.cmdline.baseclass import (VerdiCommandRouter, VerdiCommandWithSubcommands)
-from aiida.cmdline.commands import verdi, verdi_node, verdi_node_repo
-
+from aiida.cmdline.commands import verdi
+from aiida.cmdline.params import options, arguments
 from aiida.cmdline.utils import echo
 from aiida.cmdline.utils.decorators import with_dbenv
-from aiida.cmdline.params import options, arguments
 
 
-class Node(VerdiCommandRouter):
-    """
-    Manage operations on AiiDA nodes
-
-    There is a list of subcommands for managing specific types of data.
-    For instance, 'node repo' manages the files in the local repository.
-    """
-
-    def __init__(self):
-        """
-        A dictionary with valid commands and functions to be called.
-        """
-        super(Node, self).__init__()
-        self.routed_subcommands = {
-            'repo': _Repo,
-            'show': verdi,
-            'tree': verdi,
-            'delete': verdi,
-            'label': verdi,
-            'description': verdi,
-        }
+@verdi.group('node')
+def verdi_node():
+    """Inspect, create and manage nodes."""
+    pass
 
 
-# Note: this class should not be exposed directly in the main module,
-# otherwise it becomes a command of 'verdi'. Instead, we want it to be a
-# subcommand of verdi data.
-class _Repo(VerdiCommandWithSubcommands):
-    """
-    Show files and their contents in the local repository
-    """
-
-    def __init__(self):
-        """
-        A dictionary with valid commands and functions to be called.
-        """
-        super(_Repo, self).__init__()
-        self.valid_subcommands = {
-            'ls': (verdi, self.complete_none),
-            'cat': (verdi, self.complete_none),
-        }
+@verdi_node.group('repo')
+def verdi_node_repo():
+    pass
 
 
 @verdi_node_repo.command('cat')
@@ -385,250 +350,3 @@ def node_delete(nodes, follow_calls, dry_run, verbose, non_interactive):
 
     delete_nodes(
         node_pks_to_delete, follow_calls=follow_calls, dry_run=dry_run, verbosity=verbosity, force=non_interactive)
-
-
-# the classes _Label and _Description are written here,
-# but in fact they are called by the verdi calculation or verdi data
-# in fact, I don't want to allow the possibility of changing labels or
-# descriptions of codes, for that there is a separate command
-
-
-class _Label(VerdiCommandWithSubcommands):
-    """
-    See or modify the label of one or more set of nodes
-    """
-
-    def __init__(self, node_subclass='data'):
-        super(_Label, self).__init__()
-        self._node_subclass = node_subclass
-        if self._node_subclass not in ['calculation', 'data']:
-            raise ValueError("Class must be loaded with a valid node_subclass")
-
-    def _node_class_ok(self, node):
-        """Check whether node class is recognized."""
-        from aiida.orm import Calculation as OrmCalculation
-        from aiida.orm import Data as OrmData
-
-        if self._node_subclass == 'calculation':
-            return isinstance(node, OrmCalculation)
-        elif self._node_subclass == 'data':
-            return isinstance(node, OrmData)
-        else:
-            raise ValueError("node_subclass not recognized")
-
-    # pylint: disable=too-many-branches
-    def run(self, *args):
-        if not is_dbenv_loaded():
-            load_dbenv()
-        import argparse
-        from aiida.cmdline import wait_for_confirmation
-
-        parser = argparse.ArgumentParser(
-            prog=self.get_full_command_name(), description="See/modify the labels of Nodes.")
-        # display parameters
-        parser.add_argument(
-            '-r', '--raw', action='store_true', default=False, help="Display only the labels, without the pk.")
-        # pks
-        parser.add_argument('pks', type=int, nargs='+', help="a list of nodes to show.")
-        # parameters for label modification
-        parser.add_argument(
-            '-s',
-            '--set',
-            action='store_true',
-            default=False,
-            help="If present, set a new label, otherwise only "
-            "show the labels.")
-        parser.add_argument('-f', '--force', action='store_true', default=False, help="Force the reset of the label.")
-        parser.add_argument(
-            '-l',
-            '--label',
-            type=str,
-            default=None,
-            help="The new label to be set on the node. Note: "
-            "pass it between quotes.")
-
-        parsed_args = parser.parse_args(args)
-        raw = parsed_args.raw
-        # pylint: disable=invalid-name
-        pks = parsed_args.pks
-
-        if not parsed_args.set:
-            for pk in pks:
-                node = load_node(pk)
-                if not self._node_class_ok(node):
-                    print "Node {} is not a subclass of {}. Exiting...".format(pk, self._node_subclass)
-                    sys.exit(1)
-                if raw:
-                    print '"{}"'.format(node.label)
-                else:
-                    if not node.label:
-                        print 'Node {}, label: n.a.'.format(pk)
-                    else:
-                        print 'Node {}, label: "{}"'.format(pk, node.label)
-        else:
-            if len(pks) > 1:
-                sys.stderr.write("More than one node found to set one label" ". Exiting...\n")
-                sys.exit(1)
-            else:
-                pk = pks[0]
-
-            new_label = parsed_args.label
-            if new_label is None:
-                sys.stderr.write("A new label is required" ". Exiting...\n")
-                sys.exit(1)
-
-            node = load_node(pk)
-
-            if not self._node_class_ok(node):
-                print "Node {} is not a subclass of {}. Exiting...".format(pk, self._node_subclass)
-                sys.exit(1)
-
-            old_label = node.label
-            if not parsed_args.force:
-                sys.stderr.write("Current label is: {}\n".format(old_label))
-                sys.stderr.write("New label is: {}\n".format(new_label))
-                sys.stderr.write("Are you sure you want to reset the label? " "[Y/N] ")
-                if not wait_for_confirmation():
-                    sys.exit(0)
-
-            node.label = new_label
-
-
-class _Description(VerdiCommandWithSubcommands):
-    """
-    See or modify the label of one or more set of nodes
-    """
-
-    def __init__(self, node_subclass='data'):
-        super(_Description, self).__init__()
-        self._node_subclass = node_subclass
-        if self._node_subclass not in ['calculation', 'data']:
-            raise ValueError("Class must be loaded with a valid node_subclass")
-
-    def _node_class_ok(self, node):
-        """Check whether node class is recognized."""
-        from aiida.orm import Calculation as OrmCalculation
-        from aiida.orm import Data as OrmData
-
-        if self._node_subclass == 'calculation':
-            return isinstance(node, OrmCalculation)
-        elif self._node_subclass == 'data':
-            return isinstance(node, OrmData)
-        else:
-            raise ValueError("node_subclass not recognized")
-
-    # pylint: disable=too-many-statements,too-many-branches
-    def run(self, *args):
-        if not is_dbenv_loaded():
-            load_dbenv()
-        import argparse
-        from aiida.cmdline import wait_for_confirmation
-
-        parser = argparse.ArgumentParser(
-            prog=self.get_full_command_name(),
-            description="See description of Nodes. If no node "
-            "description or label is found, prints n.a.")
-
-        # parameters for display
-        parser.add_argument('-n', '--no-labels', action='store_false', default=True, help="Don't show the labels.")
-        parser.add_argument(
-            '-r',
-            '--raw',
-            action='store_true',
-            default=False,
-            help="If set, prints only the description without "
-            "pks or labels.")
-        # pks
-        parser.add_argument('pks', type=int, nargs='+', help="a list of node pks to show.")
-        # parameters for description modifications
-        parser.add_argument(
-            '-s',
-            '--set',
-            action='store_true',
-            default=False,
-            help="If present, set a new label, otherwise only "
-            "show the labels.")
-        parser.add_argument(
-            '-a',
-            '--add-to-description',
-            action='store_true',
-            default=False,
-            help="If -s, the string passed in -d is appended "
-            "to the current description.")
-        parser.add_argument(
-            '-f', '--force', action='store_true', default=False, help="Force the reset of the description.")
-        parser.add_argument(
-            '-d',
-            '--description',
-            type=str,
-            help="The new description to be set on the node. "
-            "Note: pass it between quotes.")
-
-        parsed_args = parser.parse_args(args)
-
-        # pylint: disable=invalid-name
-        pks = parsed_args.pks
-
-        if not parsed_args.set:
-            also_labels = parsed_args.no_labels
-            for pk in pks:
-                n = load_node(pk)
-
-                if not self._node_class_ok(n):
-                    echo.echo_critical("Node {} is not a subclass of {}. Exiting...".format(pk, self._node_subclass))
-
-                label = n.label
-                description = n.description
-
-                if parsed_args.raw:
-                    echo.echo('"{}"'.format(n.description))
-                    echo.echo("")
-
-                else:
-                    echo.echo("Node pk: {}".format(pk))
-                    if also_labels:
-                        if label:
-                            echo.echo('Label: "{}"'.format(label))
-                        else:
-                            echo.echo('Label: n.a.')
-                    if description:
-                        echo.echo('Description: "{}"'.format(description))
-                    else:
-                        echo.echo('Description: n.a.')
-                    echo.echo("")
-        else:
-            # check that only one pk is present
-            if len(pks) > 1:
-                sys.stderr.write("More than one node found to set one description" ". Exiting...\n")
-                sys.exit(1)
-            else:
-                pk = pks[0]
-
-            new_description = parsed_args.description
-            if new_description is None:
-                sys.stderr.write("No description was found. Exiting...\n")
-                sys.exit(1)
-
-            if not parsed_args.add_to_description:
-                n = load_node(pk)
-                if not self._node_class_ok(n):
-                    echo.echo_critical("Node {} is not a subclass of {}. Exiting...".format(pk, self._node_subclass))
-
-                old_description = n.description
-                if not parsed_args.force:
-                    sys.stderr.write("Current description is: {}\n".format(old_description))
-                    sys.stderr.write("New description is: {}\n".format(new_description))
-                    sys.stderr.write("Are you sure you want to reset the description? " "[Y/N] ")
-                    if not wait_for_confirmation():
-                        sys.exit(0)
-
-                n.description = new_description
-
-            else:
-                n = load_node(pk)
-                if not self._node_class_ok(n):
-                    echo.echo_critical("Node {} is not a subclass of {}. Exiting...".format(pk, self._node_subclass))
-
-                old_description = n.description
-                new_description = old_description + "\n" + new_description
-                n.description = new_description
